requirements.txt
djangorestframework==3.10.0
djangorestframework-jwt==1.11.0
django==2.2.3
mongoengine==0.18.2
mysqlclient==1.4.2.post1
requests==2.22.0
redis==3.2.1
django_redis==4.10.0
django-cors-headers==3.0.2
PySnooper==0.2.2
#git+https://github.com/philipn/django-rest-framework-filters.git
djangorestframework-filters==1.0.0.dev0
ipython==7.6.1
Pillow==6.1.0
django-aliyun-oss2-storage==0.1.5
django-extensions==2.2.1
celery==4.3.0
django-celery-beat==1.5.0